+++ wnb/templates/wandb-cicd-project/components/__init__.py
@@ -1,0 +1,1 @@
+# Components package for wandb-cicd-project